services/service_daemon/requirements.txt
kafka-python-ng
schedule
python-dotenv
pymysql
cryptography
googlemaps
google-api-python-client
google-auth-oauthlib